Backend_requirements.txt
numpy
nltk
tensorflow
flask
flask-cors
flask-session
pymongo
nltk
tensorflow
werkzeug
pyjwt
python-dotenv
torch
transformers
scikit-learn
sentencepiece
